requirements.txt
pandas
fastapi
uvicorn
python-dotenv
python-multipart
jinja2
langchain
langchain_core
langchain_community
langchain_google_genai
langchain_groq
langchain_openai
langchain_pinecone
langchain_chroma
langchain_tavily
fitz
langgraph
pinecone
langchain_text_splitters
streamlit